pyproject.toml
[tool.poetry]
name = "ai-job-screener"
version = "0.1.0"
description = "AI-powered job screener using SerpAPI and OpenAI to rank jobs and generate cover letters."
authors = ["Varan Singh Rohila <varansinghrohila@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.3.0"
python-dotenv = "^1.1.0"
openai = "^1.88.0"
tqdm = "^4.67.1"
python-docx = "^1.2.0"
google-search-results = "^2.4.2"
openpyxl = "^3.1.5"
ipykernel = "^6.29.5"
commitizen = "^4.8.3"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pytest = "^8.2.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"